requirements.txt
altgraph==0.17.4
appdirs==1.4.4
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
fake-useragent==2.0.3
filelock==3.17.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
lxml_html_clean==0.4.1
msedge-selenium-tools==3.141.4
outcome==1.3.0.post0
packaging==24.1
parse==1.20.2
parsel==1.10.0
pefile==2023.2.7
pluggy==1.5.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==11.1.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyOpenSSL==25.0.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
pytest==8.3.3
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
queuelib==1.7.0
Random-Word==1.0.11
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
selenium==4.29.0
selenium-stealth==1.0.6
service-identity==24.2.0
setuptools==75.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
telebot==0.0.5
tldextract==5.1.3
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==1.26.20
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
zope.interface==7.2
